Style changes and some refactoring
